feature(#191): add splash screen with risingbrain logo
