+++ Android/app/src/main/AndroidManifest.xml
@@ -39,18 +39,20 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".ui.setting.view.SettingActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.chat.view.ChatActivity"
-            android:exported="true"
-            android:windowSoftInputMode="adjustResize">
+            android:name=".ui.splash.SplashActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".ui.setting.view.SettingActivity"
+            android:exported="false" />
+        <activity
+            android:name=".ui.chat.view.ChatActivity"
+            android:exported="false" />
 
         <receiver
             android:name=".utils.helpers.chat.AlarmReceiver"
